Backend/requirements.txt
fastapi
uvicorn
google-cloud-storage
tensorflow-cpu
numpy
pandas